Convert all BRL-CAD magic numbers to uint32_t.  This change affects more than a dozen structs and more than 500 usages throughout the code including many function signatures, but is still considered minimally impacting.  This consolidates the slew of int, long, unsigned int, unsigned long, etc definitions of magic numbers that were assuming to be 4 bytes for magic number validation.  It also exposed a curious book-keeping practice in NMG where pointers to the magic number are used for tracking topology and NMG operations.  Now the intent is explicit and matches the data sizes defined via libbu's magic number checking interface, as well as elsewhere throughout the code.  The magic number struct values are now fixed in size regardless of 32-bit or 64-bit compilation and preliminary regression testing indicates no change in behavior.
svn:revision:45858
svn:branch:trunk
svn:account:brlcad

+++ include/analyze.h
@@ -54,7 +54,7 @@ __BEGIN_DECLS
 
 /* the entries in the density table */
 struct density_entry {
-    long magic;
+    uint32_t magic;
     double grams_per_cu_mm;
     char *name;
 };
